+++ druid/src/lib.rs
@@ -172,7 +172,7 @@ pub use shell::{
     Application, Clipboard, ClipboardFormat, Code, Cursor, Error as PlatformError,
     FileDialogOptions, FileInfo, FileSpec, FormatId, HotKey, KbKey, KeyEvent, Location, Modifiers,
     Monitor, MouseButton, MouseButtons, RawMods, Scalable, Scale, Screen, SysMods, Text,
-    TimerToken, WindowHandle,
+    TimerToken, WindowHandle, WindowState,
 };
 
 pub use crate::core::WidgetPod;